コンフリクト解消、商品詳細画面の実装
